ai_project/requirements.txt
SpeechRecognition>=3.8.1
pyttsx3>=2.90
openai>=1.0.0
python-dotenv>=1.0.0
